pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trend_follower_xauusd"
version = "0.1.0"
dependencies = [
  "pandas>=2.0",
  "PyYAML>=6.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
